+++ spec/support/have_styling_matcher.rb
@@ -13,7 +13,7 @@ RSpec::Matchers.define :have_styling do |rules|
     end
   end
 
-  describe { "have styles #{rules.inspect} at selector #{@selector.inspect}" }
+  description { "have styles #{rules.inspect} at selector #{@selector.inspect}" }
   failure_message_for_should { |document| "expected styles at #{@selector.inspect} to be #{rules.inspect} but was #{parsed_styles(document).inspect}" }
   failure_message_for_should_not { "expected styles at #{@selector.inspect} to not be #{rules.inspect}" }