+++ spec/support/parse_styling.rb
@@ -9,9 +9,17 @@ module SpecHelpers
     end
 
     def parse_styling(styles)
-      styles.split(';').inject([]) do |array, item|
-        array << item.split(':', 2).map(&:strip)
-      end
+      styles.split(';').map { |style| parse_style(style) }
     end
+
+    private
+      def parse_style(style)
+        rule, value = style.split(':', 2).map(&:strip)
+        [rule, normalize_escaped_quotes(value)]
+      end
+
+      def normalize_escaped_quotes(string)
+        string.gsub('%22', '"')
+      end
   end
 end