+++ Changelog.md
@@ -6,6 +6,9 @@
   * Add `config.roadie.enabled` that can be set to `false` to disable Roadie completely.
 * Bug fixes:
   * Proc objects to the `:css` option is now run in the context of the mailer instance, mirroring similar options from ActionMailer.
+  * Fix some tests that would always pass
+  * Improve JRuby compatibility
+  * Update Gemfile.lock and fix issues with newer gem versions
 
 ### 2.3.3